[YAML] Automatically add simulated clusters from the clusters directory into the list of clusters used for test generation instead of manually adding them one by one at the top of SimulatedClusters.js
